Extract Trainer test and partitions divisions into the separate test
