+++ modules/ml/src/test/java/org/apache/ignite/ml/clustering/KMeansTrainerTest.java
@@ -45,12 +45,12 @@ public class KMeansTrainerTest {
     @Test
     public void findOneClusters() {
         Map<Integer, double[]> data = new HashMap<>();
-        data.put(0, new double[] {1.0, 1.0, 1.0});
-        data.put(1, new double[] {1.0, 2.0, 1.0});
-        data.put(2, new double[] {2.0, 1.0, 1.0});
-        data.put(3, new double[] {-1.0, -1.0, 2.0});
-        data.put(4, new double[] {-1.0, -2.0, 2.0});
-        data.put(5, new double[] {-2.0, -1.0, 2.0});
+        data.put(0, new double[]{1.0, 1.0, 1.0});
+        data.put(1, new double[]{1.0, 2.0, 1.0});
+        data.put(2, new double[]{2.0, 1.0, 1.0});
+        data.put(3, new double[]{-1.0, -1.0, 2.0});
+        data.put(4, new double[]{-1.0, -2.0, 2.0});
+        data.put(5, new double[]{-2.0, -1.0, 2.0});
 
         KMeansTrainer trainer = new KMeansTrainer()
             .withDistance(new EuclideanDistance())
@@ -62,15 +62,17 @@ public class KMeansTrainerTest {
         assertEquals(2, trainer.getSeed());
         assertTrue(trainer.getDistance() instanceof EuclideanDistance);
 
-        KMeansModel knnMdl = trainer.withK(1).fit(
-            new LocalDatasetBuilder<>(data, 2),
-            (k, v) -> VectorUtils.of(Arrays.copyOfRange(v, 0, v.length - 1)),
-            (k, v) -> v[2]
-        );
+        KMeansModel knnMdl = trainer
+            .withK(1)
+            .fit(
+                new LocalDatasetBuilder<>(data, 2),
+                (k, v) -> VectorUtils.of(Arrays.copyOfRange(v, 0, v.length - 1)),
+                (k, v) -> v[2]
+            );
 
-        Vector firstVector = new DenseVector(new double[] {2.0, 2.0});
+        Vector firstVector = new DenseVector(new double[]{2.0, 2.0});
         assertEquals(knnMdl.apply(firstVector), 0.0, PRECISION);
-        Vector secondVector = new DenseVector(new double[] {-2.0, -2.0});
+        Vector secondVector = new DenseVector(new double[]{-2.0, -2.0});
         assertEquals(knnMdl.apply(secondVector), 0.0, PRECISION);
         assertEquals(trainer.getMaxIterations(), 1);
         assertEquals(trainer.getEpsilon(), PRECISION, PRECISION);
