+++ modules/ml/src/test/java/org/apache/ignite/ml/regressions/logistic/LogRegMultiClassTrainerTest.java
@@ -20,47 +20,22 @@ package org.apache.ignite.ml.regressions.logistic;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ThreadLocalRandom;
 import org.apache.ignite.ml.TestUtils;
+import org.apache.ignite.ml.common.TrainerTest;
 import org.apache.ignite.ml.math.primitives.vector.VectorUtils;
-import org.apache.ignite.ml.math.primitives.vector.impl.DenseVector;
 import org.apache.ignite.ml.nn.UpdatesStrategy;
 import org.apache.ignite.ml.optimization.SmoothParametrized;
 import org.apache.ignite.ml.optimization.updatecalculators.SimpleGDParameterUpdate;
 import org.apache.ignite.ml.optimization.updatecalculators.SimpleGDUpdateCalculator;
 import org.apache.ignite.ml.regressions.logistic.multiclass.LogRegressionMultiClassModel;
 import org.apache.ignite.ml.regressions.logistic.multiclass.LogRegressionMultiClassTrainer;
-import org.apache.ignite.ml.svm.SVMLinearBinaryClassificationTrainer;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 /**
  * Tests for {@link LogRegressionMultiClassTrainer}.
  */
-@RunWith(Parameterized.class)
-public class LogRegMultiClassTrainerTest {
-    /** Parameters. */
-    @Parameterized.Parameters(name = "Data divided on {0} partitions")
-    public static Iterable<Integer[]> data() {
-        return Arrays.asList(
-            new Integer[]{1},
-            new Integer[]{2},
-            new Integer[]{3},
-            new Integer[]{5},
-            new Integer[]{7},
-            new Integer[]{100}
-        );
-    }
-
-    /** Number of partitions. */
-    @Parameterized.Parameter
-    public int parts;
-
-    /** Precision in test checks. */
-    private static final double PRECISION = 1e-2;
-
+public class LogRegMultiClassTrainerTest extends TrainerTest {
     /**
      * Test trainer on 4 sets grouped around of square vertices.
      */
@@ -68,8 +43,8 @@ public class LogRegMultiClassTrainerTest {
     public void testTrainWithTheLinearlySeparableCase() {
         Map<Integer, double[]> cacheMock = new HashMap<>();
 
-        for (int i = 0; i < data.length; i++)
-            cacheMock.put(i, data[i]);
+        for (int i = 0; i < fourSetsInSquareVertices.length; i++)
+            cacheMock.put(i, fourSetsInSquareVertices[i]);
 
         final UpdatesStrategy<SmoothParametrized, SimpleGDParameterUpdate> stgy = new UpdatesStrategy<>(
             new SimpleGDUpdateCalculator(0.2),
@@ -106,407 +81,4 @@ public class LogRegMultiClassTrainerTest {
         TestUtils.assertEquals(2, mdl.apply(VectorUtils.of(-10, -10)), PRECISION);
         TestUtils.assertEquals(3, mdl.apply(VectorUtils.of(10, -10)), PRECISION);
     }
-    /** Two-easy clustered data by y = x. */
-    private static final double[][] data = {
-        {0, 9.35096604945605, 9.946073797069054},
-        {0, 9.135109633114403, 9.962676066205383},
-        {0, 9.046654725589521, 9.610699793950662},
-        {0, 9.827221553421282, 9.4176319880153},
-        {0, 9.277441430833566, 9.502990699976},
-        {0, 9.444827307967367, 9.903310367805602},
-        {0, 9.911404997680545, 9.226246217883297},
-        {0, 9.950231642973769, 9.453518533258803},
-        {0, 9.281545278543017, 9.438272102773379},
-        {0, 9.032306746555102, 9.517675092676706},
-        {0, 9.286542956290456, 9.15288903978334},
-        {0, 9.896451632473255, 9.019751070009821},
-        {0, 9.611642481367562, 9.17209652044495},
-        {0, 9.592540623266126, 9.306160678545629},
-        {0, 9.817470117880873, 9.838651444371973},
-        {0, 9.263220850397941, 9.139179322873582},
-        {0, 9.949097640181272, 9.624710378790242},
-        {0, 9.616004097319287, 9.421557303733453},
-        {0, 9.512900976289933, 9.28642137092367},
-        {0, 9.207793663546337, 9.40094289636865},
-        {0, 9.079279410265883, 9.76978559451163},
-        {0, 9.328945661288095, 9.645773710532888},
-        {0, 9.80101696222916, 9.511903913501255},
-        {0, 9.882593127029741, 9.73545127073394},
-        {0, 9.75372887212885, 9.435141350132769},
-        {0, 9.288527674365598, 9.055665753045206},
-        {0, 9.88272159816372, 9.055932205550423},
-        {0, 9.385642321423624, 9.922172934733265},
-        {0, 9.830217517055729, 9.415174260405154},
-        {0, 9.184970761195489, 9.03515483431538},
-        {0, 9.747503155479809, 9.38708759338332},
-        {0, 9.953962908254736, 9.483949174467012},
-        {0, 9.271685731881993, 9.128890010491494},
-        {0, 9.441240324686845, 9.07960435205457},
-        {0, 9.168560731741703, 9.256530860101089},
-        {0, 9.010517147230432, 9.94335328515589},
-        {0, 9.1749227239244, 9.018681913631386},
-        {0, 9.413360501729251, 9.302212703700196},
-        {0, 9.439461439481182, 9.318631395882242},
-        {0, 9.531551691985907, 9.232525664308465},
-        {0, 9.466805772615563, 9.511711890834333},
-        {0, 9.633242901042053, 9.972778102570045},
-        {0, 9.517692290376388, 9.73537462150143},
-        {0, 9.187046049036134, 9.059073377533783},
-        {0, 9.121523234392956, 9.504221886903101},
-        {0, 9.493957951674021, 9.608201135992367},
-        {0, 9.981993764415321, 9.333278989889811},
-        {0, 9.371277571698762, 9.110041365023866},
-        {0, 9.681446270907697, 9.7870063720198},
-        {0, 9.639466883264246, 9.434768030033164},
-        {0, 9.391982858267035, 9.934707093985823},
-        {0, 9.550060071547726, 9.473132681990514},
-        {0, 9.256562054384402, 9.211913854106896},
-        {0, 9.46408385327689, 9.158869250798142},
-        {0, 9.442994981367162, 9.189227375629654},
-        {0, 9.697833866121318, 9.21112449845501},
-        {0, 9.115534908153043, 9.115227178046245},
-        {0, 9.835218474137239, 9.98174155822633},
-        {0, 9.026698146309743, 9.248759846540965},
-        {0, 9.68118581769866, 9.40512628823504},
-        {0, 9.81721640069966, 9.369105145483651},
-        {0, 9.975877208452287, 9.640693828024975},
-        {0, 9.823272242807437, 9.46823993908653},
-        {0, 9.638281188176519, 9.534774307683374},
-        {0, 9.597003178481613, 9.84238115941204},
-        {0, 9.941999007792681, 9.331877359355289},
-        {0, 9.050540877852525, 9.244472301490417},
-        {0, 9.358931306187054, 9.900809398285286},
-        {0, 9.170247599517836, 9.87585551194908},
-        {0, 9.461705027907554, 9.167319400226486},
-        {0, 9.076729207165052, 9.677578134220534},
-        {0, 9.488544686081216, 9.62380634923249},
-        {0, 9.929150661994122, 9.152491122614597},
-        {0, 9.890051482992417, 9.1709621079536},
-        {0, 9.839485513056095, 9.643849781319778},
-        {0, 9.749461922180853, 9.045432748127462},
-        {0, 9.58439542919333, 9.225044809549836},
-        {0, 9.479465134364697, 9.706551666966702},
-        {0, 9.00707492076871, 9.839317970534172},
-        {0, 9.948409701102793, 9.380261430658763},
-        {0, 9.264850115578076, 9.696516344063658},
-        {0, 9.977078194073387, 9.213405339955512},
-        {0, 9.648087669569941, 9.898977891084664},
-        {0, 9.724090075117749, 9.876133066062916},
-        {0, 9.445249316659568, 9.373023119966643},
-        {0, 9.995541563884071, 9.57923804140667},
-        {0, 9.667359233860397, 9.720098746660245},
-        {0, 9.379303845088474, 9.520602789251743},
-        {0, 9.996287800651865, 9.838061655335768},
-        {0, 9.318835567328465, 9.009915558605616},
-        {0, 9.103894679089793, 9.674971708485224},
-        {0, 9.346826400314828, 9.888779618232787},
-        {0, 9.659116962016478, 9.608712473271416},
-        {0, 9.661516337354719, 9.416786365864226},
-        {0, 9.642593770590324, 9.251344999039574},
-        {0, 9.134003475979116, 9.551760245909657},
-        {0, 9.524862003327057, 9.307789887454172},
-        {0, 9.883705581666579, 9.325086464359684},
-        {0, 9.96076863440133, 9.81636527085299},
-        {0, 9.995704158311584, 9.544553004819253},
-        {1, -9.094953387232211, 9.06233128328723},
-        {1, -9.304897363378368, 9.143926554861004},
-        {1, -9.03524958020074, 9.370326522034881},
-        {1, -9.120893310395626, 9.271851530835537},
-        {1, -9.510902040922451, 9.2470398948938},
-        {1, -9.6525973741057, 9.725355730393005},
-        {1, -9.65730261326345, 9.757814601272596},
-        {1, -9.597463454487615, 9.870093256106818},
-        {1, -9.190101362739775, 9.594505054154807},
-        {1, -9.72020516663928, 9.49084494643775},
-        {1, -9.723347588431338, 9.129139508430457},
-        {1, -9.33996314024198, 9.525934956132764},
-        {1, -9.824803485424123, 9.128546700002982},
-        {1, -9.346973220919576, 9.934992542662958},
-        {1, -9.685940369418338, 9.30810392592615},
-        {1, -9.064058121381708, 9.846942888423445},
-        {1, -9.368987058951426, 9.557135466015499},
-        {1, -9.782353308524383, 9.857550405413855},
-        {1, -9.281500887267686, 9.056968941046172},
-        {1, -9.514451522447168, 9.713696846961527},
-        {1, -9.607099689382135, 9.682075033940093},
-        {1, -9.144871412854759, 9.146320338346246},
-        {1, -9.54203309158306, 9.220014377847022},
-        {1, -9.238269645840251, 9.948063795512258},
-        {1, -9.286942806777112, 9.522342489392214},
-        {1, -9.591474157985536, 9.240285207594253},
-        {1, -9.652843973116592, 9.557983695755953},
-        {1, -9.126794849562028, 9.452966323026885},
-        {1, -9.877221229728452, 9.151312939643672},
-        {1, -9.170379066479606, 9.381576400806694},
-        {1, -9.411298671068392, 9.133322302544746},
-        {1, -9.666443924685849, 9.66428867311317},
-        {1, -9.347964494643556, 9.012849397302583},
-        {1, -9.493681117964078, 9.332240464982554},
-        {1, -9.623975723800413, 9.419921503264844},
-        {1, -9.292219487063763, 9.00214102314859},
-        {1, -9.194419464738496, 9.640048387436925},
-        {1, -9.886720923292938, 9.834939723803704},
-        {1, -9.90520284610924, 9.17595267606471},
-        {1, -9.284829868633738, 9.268795876426012},
-        {1, -9.498878372098952, 9.5997098342015},
-        {1, -9.359302922869169, 9.47880701571168},
-        {1, -9.258562740082393, 9.497531680793207},
-        {1, -9.895388929537848, 9.00756585816333},
-        {1, -9.627928477333924, 9.391262771761872},
-        {1, -9.525281129279826, 9.796892255719904},
-        {1, -9.59598592778135, 9.067874949457092},
-        {1, -9.110283105135892, 9.821653780489235},
-        {1, -9.343973780672988, 9.63557812382392},
-        {1, -9.87812414314095, 9.978007969979139},
-        {1, -9.98832246915748, 9.623150872300222},
-        {1, -9.115997082508613, 9.965470531748467},
-        {1, -9.874391718339105, 9.214113577543877},
-        {1, -9.671664494678888, 9.15862012290195},
-        {1, -9.031596433460688, 9.616814958480965},
-        {1, -9.758627761132653, 9.511908952613643},
-        {1, -9.205087108977219, 9.840949306240816},
-        {1, -9.171734592697309, 9.702842939318314},
-        {1, -9.082886085070493, 9.524201651321903},
-        {1, -9.74595864484071, 9.219346103723025},
-        {1, -9.898468941378516, 9.994402484197503},
-        {1, -9.341582531784448, 9.193680038418634},
-        {1, -9.570090524257228, 9.201198104295603},
-        {1, -9.88361320124743, 9.027615263347323},
-        {1, -9.154222720481965, 9.799927021695417},
-        {1, -9.364221227791875, 9.042090834574182},
-        {1, -9.333131749015948, 9.790442620484125},
-        {1, -9.286700941581561, 9.89073867458494},
-        {1, -9.348737197099151, 9.637939060929087},
-        {1, -9.442420524656606, 9.07802294456236},
-        {1, -9.069329135123306, 9.658515489139848},
-        {1, -9.306682910312364, 9.20831776028291},
-        {1, -9.033846541544232, 9.32904963306478},
-        {1, -9.706767953982897, 9.9204656840812},
-        {1, -9.855922299233484, 9.212398390928783},
-        {1, -9.31778377138365, 9.001381041592891},
-        {1, -9.498262395904716, 9.627240779587641},
-        {1, -9.165515191167106, 9.8269942856602},
-        {1, -9.975445549855277, 9.940934989111799},
-        {1, -9.083105286998059, 9.006127740460453},
-        {1, -9.570145038082837, 9.682155599203648},
-        {1, -9.61392195996382, 9.417864984298848},
-        {1, -9.274771331302999, 9.641773516631659},
-        {1, -9.296296304670749, 9.782496135034126},
-        {1, -9.906415110246952, 9.754391405446135},
-        {1, -9.401887484923442, 9.177845637020802},
-        {1, -9.530971211940608, 9.165119804525942},
-        {1, -9.82379861350907, 9.79567065636976},
-        {1, -9.652776399686564, 9.905939382705197},
-        {1, -9.876593047451918, 9.945310791455892},
-        {1, -9.663611565135188, 9.362793091580434},
-        {1, -9.199103361444621, 9.635196006461447},
-        {1, -9.190013322848332, 9.124127000468004},
-        {1, -9.29736354578434, 9.717999298890678},
-        {1, -9.220547853711237, 9.559927412569595},
-        {1, -9.300431356958706, 9.76396216541998},
-        {1, -9.157649670754807, 9.990846988919046},
-        {1, -9.681918677002109, 9.68618286595764},
-        {1, -9.309195235661146, 9.312880801021818},
-        {1, -9.061160475710913, 9.076614202325946},
-        {2, -9.062489260904384, -9.29639290758419},
-        {2, -9.228543182338143, -9.678377216077045},
-        {2, -9.058090832908235, -9.193945883550121},
-        {2, -9.133051729493113, -9.591373007767894},
-        {2, -9.287844094445548, -9.551255004015},
-        {2, -9.007505358739156, -9.364102496975889},
-        {2, -9.573448348548297, -9.721351111009751},
-        {2, -9.839063104064442, -9.913376420693114},
-        {2, -9.009615911555375, -9.726047024128608},
-        {2, -9.101017317976435, -9.704243867142955},
-        {2, -9.982108914119253, -9.16651010251761},
-        {2, -9.446194150458751, -9.254956921695555},
-        {2, -9.189473272816354, -9.810681137049205},
-        {2, -9.118077427599777, -9.540094810610913},
-        {2, -9.771250464767986, -9.523914718655663},
-        {2, -9.66962428717098, -9.363171620624835},
-        {2, -9.312167530669402, -9.343252976723711},
-        {2, -9.464223946364095, -9.030677424225916},
-        {2, -9.360044171938823, -9.307110078788382},
-        {2, -9.011501658023455, -9.36530250968901},
-        {2, -9.775885771959839, -9.99889314514122},
-        {2, -9.674611861667914, -9.258187855592231},
-        {2, -9.738640777018995, -9.111785670315005},
-        {2, -9.246690988432968, -9.721028941948624},
-        {2, -9.390261807995243, -9.588861735182837},
-        {2, -9.291113352727827, -9.269267155328981},
-        {2, -9.851335630543913, -9.706611637556188},
-        {2, -9.585157995064394, -9.405552049981731},
-        {2, -9.918436572526948, -9.16760421314763},
-        {2, -9.845493743518675, -9.355482956823167},
-        {2, -9.731220848845956, -9.225343258111073},
-        {2, -9.222705334863235, -9.494812693860784},
-        {2, -9.981016698450784, -9.905493543993186},
-        {2, -9.46735837748333, -9.4826245649012},
-        {2, -9.244414641225871, -9.747631208358092},
-        {2, -9.055383358563462, -9.531078141057671},
-        {2, -9.769432919539609, -9.61352749756392},
-        {2, -9.369900693663043, -9.108143337018905},
-        {2, -9.607028386780009, -9.114073621581822},
-        {2, -9.777391839524553, -9.011542490337462},
-        {2, -9.006992341646022, -9.807142738339437},
-        {2, -9.268800709859363, -9.64049865255139},
-        {2, -9.675247117678266, -9.59986912340877},
-        {2, -9.64637138569114, -9.373492063216789},
-        {2, -9.107966171477159, -9.89296888054194},
-        {2, -9.844813041035149, -9.265286980535892},
-        {2, -9.741557572466677, -9.332262102684087},
-        {2, -9.877113842998332, -9.236779890169021},
-        {2, -9.717067250147496, -9.064661056318842},
-        {2, -9.621588376526242, -9.877688077281952},
-        {2, -9.517814042484112, -9.540587835450802},
-        {2, -9.301056957528804, -9.825047942369075},
-        {2, -9.571496535251406, -9.7886645523611},
-        {2, -9.720509286872675, -9.391715190333258},
-        {2, -9.440573147395899, -9.788983529514448},
-        {2, -9.26187156355727, -9.6495064067468},
-        {2, -9.658496105019307, -9.56612823492413},
-        {2, -9.380443710902496, -9.68085867523561},
-        {2, -9.942337341048844, -9.051311192273833},
-        {2, -9.078217384202866, -9.916249447505033},
-        {2, -9.040049870218203, -9.034931091928817},
-        {2, -9.38309299369458, -9.652061835126116},
-        {2, -9.2185666133056, -9.230952753648268},
-        {2, -9.556045604713985, -9.68622417688499},
-        {2, -9.763408055045888, -9.879577482698977},
-        {2, -9.4247203087675, -9.639176769093654},
-        {2, -9.87358328609414, -9.895570735983382},
-        {2, -9.819441742886346, -9.8365613475581},
-        {2, -9.658089225310723, -9.489731885421016},
-        {2, -9.943990436893316, -9.452660725226394},
-        {2, -9.499576083220616, -9.936796975306573},
-        {2, -9.209278737078256, -9.515912871664437},
-        {2, -9.822627739746856, -9.208467277950026},
-        {2, -9.250697491903084, -9.388580147580788},
-        {2, -9.499425743259364, -9.350980719673753},
-        {2, -9.275926339651928, -9.617104986484284},
-        {2, -9.1796228747286, -9.600489090237376},
-        {2, -9.349551823375743, -9.006466892950566},
-        {2, -9.894633921415739, -9.68766261225829},
-        {2, -9.65858550958029, -9.981852526887},
-        {2, -9.260496691277194, -9.809097777192473},
-        {2, -9.819512412109138, -9.897278497873733},
-        {2, -9.095722203640902, -9.373361177409254},
-        {2, -9.350211015838992, -9.8070103689666},
-        {2, -9.666932714082296, -9.012476306081684},
-        {2, -9.808494394881976, -9.141856503904373},
-        {2, -9.659369482494562, -9.813220865668578},
-        {2, -9.695328684452264, -9.348824074007899},
-        {2, -9.559852026507784, -9.874175917293163},
-        {2, -9.09372192117967, -9.409697201374975},
-        {2, -9.287303427948462, -9.501710345463191},
-        {2, -9.559530752361578, -9.915461534620048},
-        {2, -9.582664553428488, -9.403076102016477},
-        {2, -9.52173572568699, -9.741375773070464},
-        {2, -9.65354706029232, -9.818082622224445},
-        {2, -9.415838021477068, -9.532580879297706},
-        {2, -9.574004758496413, -9.086286237660188},
-        {2, -9.270611925252807, -9.702167164438746},
-        {2, -9.95686463396123, -9.162427711243494},
-        {2, -9.274599236711888, -9.877754856511778},
-        {3, 9.14689232210878, -9.775341371070157},
-        {3, 9.023355885230728, -9.650091265039629},
-        {3, 9.445914402990603, -9.367844134021585},
-        {3, 9.12739637867819, -9.790557561179597},
-        {3, 9.041303217790349, -9.586261899702581},
-        {3, 9.205210383417626, -9.985844424729768},
-        {3, 9.760747183322884, -9.739749414833623},
-        {3, 9.43601987318095, -9.299718258558077},
-        {3, 9.14356842517825, -9.295462642558103},
-        {3, 9.833809705258039, -9.468846417300268},
-        {3, 9.718477547371677, -9.823866211360837},
-        {3, 9.922658697442182, -9.769889056711964},
-        {3, 9.33663363137869, -9.088267105115708},
-        {3, 9.137230799593524, -9.764401780467223},
-        {3, 9.858088980083506, -9.304992329702712},
-        {3, 9.382828855133841, -9.493306421449871},
-        {3, 9.798884510277261, -9.380868512833228},
-        {3, 9.62129491417874, -9.539240839001467},
-        {3, 9.779444510688629, -9.192918853610157},
-        {3, 9.200804425227417, -9.214343851278091},
-        {3, 9.775531213188497, -9.073023597174036},
-        {3, 9.390609731389022, -9.977531450420052},
-        {3, 9.852766749781729, -9.994823748876888},
-        {3, 9.206238360247045, -9.67091791978384},
-        {3, 9.188602950870685, -9.510463637602879},
-        {3, 9.331589555754434, -9.922823935052168},
-        {3, 9.476697182752012, -9.313064140116326},
-        {3, 9.356805613304504, -9.835977587036306},
-        {3, 9.339818380404573, -9.255810669018475},
-        {3, 9.17366847248557, -9.947584334056048},
-        {3, 9.53360390823212, -9.795041609984915},
-        {3, 9.609560038477422, -9.285015745600694},
-        {3, 9.577553857280723, -9.96914900300197},
-        {3, 9.464374595524664, -9.618239089480822},
-        {3, 9.398719356212853, -9.7406758194444},
-        {3, 9.154688949078198, -9.248998548314239},
-        {3, 9.679073636776373, -9.965328464852867},
-        {3, 9.47893626848198, -9.9671543632786},
-        {3, 9.068547258387513, -9.297377035663166},
-        {3, 9.076923603177063, -9.914463831030272},
-        {3, 9.976578331543791, -9.360722370503666},
-        {3, 9.089452654960278, -9.675962954595512},
-        {3, 9.070526769096297, -9.878206691195288},
-        {3, 9.930847945955737, -9.07583308430197},
-        {3, 9.241217613699337, -9.631175172125698},
-        {3, 9.124100921554351, -9.228953372107389},
-        {3, 9.508344880276217, -9.860603437908713},
-        {3, 9.11156100183317, -9.325392997885503},
-        {3, 9.817235693989044, -9.39425968469714},
-        {3, 9.001600449220064, -9.425174755596974},
-        {3, 9.548114105927628, -9.808330723888258},
-        {3, 9.26226050324015, -9.767116578977086},
-        {3, 9.614597629315545, -9.041844364395292},
-        {3, 9.538354218499835, -9.098393947752555},
-        {3, 9.103392813936214, -9.09952673162608},
-        {3, 9.420097750306217, -9.098700662928707},
-        {3, 9.751668557712422, -9.38794903932924},
-        {3, 9.931275926738792, -9.567768498966414},
-        {3, 9.046080675655736, -9.638494792341994},
-        {3, 9.770315794108765, -9.43037261292599},
-        {3, 9.752980345824852, -9.748629501818872},
-        {3, 9.451157497026747, -9.122294173303064},
-        {3, 9.8842318143802, -9.26500677925286},
-        {3, 9.757618739984443, -9.43137249310142},
-        {3, 9.312628300108653, -9.35390228978602},
-        {3, 9.290443903557156, -9.235565486135597},
-        {3, 9.006123561818931, -9.152783217337547},
-        {3, 9.570084759165916, -9.927282503148907},
-        {3, 9.421900208122063, -9.081045753111953},
-        {3, 9.653736596553786, -9.901709124803725},
-        {3, 9.18417654510616, -9.251983632346962},
-        {3, 9.528620521688604, -9.153806541933662},
-        {3, 9.804333603959915, -9.140503586471738},
-        {3, 9.450969957775413, -9.158071229394206},
-        {3, 9.20163405176059, -9.485982651544383},
-        {3, 9.54780101021909, -9.037382999154193},
-        {3, 9.075028540176401, -9.398829949196564},
-        {3, 9.874199751417624, -9.811808331246274},
-        {3, 9.954958362231842, -9.233648957978966},
-        {3, 9.03258466527879, -9.432021155003355},
-        {3, 9.611337142970585, -9.18050106929956},
-        {3, 9.748186934551581, -9.283395815931486},
-        {3, 9.203357880317562, -9.734448423320636},
-        {3, 9.918736141570848, -9.139927237002235},
-        {3, 9.923607379931829, -9.747509729243488},
-        {3, 9.11966639233584, -9.144799648581206},
-        {3, 9.332910738465808, -9.836313230806809},
-        {3, 9.72782406722705, -9.636979470475252},
-        {3, 9.587733884348484, -9.429388313887008},
-        {3, 9.713001308076503, -9.378155762534723},
-        {3, 9.553879064305177, -9.456119811781296},
-        {3, 9.326716553614768, -9.398273985573583},
-        {3, 9.845721054911849, -9.2609414976378},
-        {3, 9.43826634715752, -9.226109072709436},
-        {3, 9.46721793264904, -9.959943210987339},
-        {3, 9.47560676057465, -9.963482009295927},
-        {3, 9.006435968586619, -9.202759792205478},
-        {3, 9.053062605095485, -9.798289703474865},
-        {3, 9.959296741639132, -9.762961500922069},
-        {3, 9.882357321966778, -9.069477551120192}
-    };
 }
