+++ modules/ml/src/test/java/org/apache/ignite/ml/regressions/logistic/LogisticRegressionSGDTrainerTest.java
@@ -20,44 +20,20 @@ package org.apache.ignite.ml.regressions.logistic;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ThreadLocalRandom;
 import org.apache.ignite.ml.TestUtils;
+import org.apache.ignite.ml.common.TrainerTest;
 import org.apache.ignite.ml.math.primitives.vector.VectorUtils;
-import org.apache.ignite.ml.math.primitives.vector.impl.DenseVector;
 import org.apache.ignite.ml.nn.UpdatesStrategy;
 import org.apache.ignite.ml.optimization.updatecalculators.SimpleGDParameterUpdate;
 import org.apache.ignite.ml.optimization.updatecalculators.SimpleGDUpdateCalculator;
 import org.apache.ignite.ml.regressions.logistic.binomial.LogisticRegressionModel;
 import org.apache.ignite.ml.regressions.logistic.binomial.LogisticRegressionSGDTrainer;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 /**
  * Tests for {@link LogisticRegressionSGDTrainer}.
  */
-@RunWith(Parameterized.class)
-public class LogisticRegressionSGDTrainerTest {
-    /** Precision in test checks. */
-    private static final double PRECISION = 1e-2;
-
-    /** Parameters. */
-    @Parameterized.Parameters(name = "Data divided on {0} partitions")
-    public static Iterable<Integer[]> data() {
-        return Arrays.asList(
-            new Integer[]{1},
-            new Integer[]{2},
-            new Integer[]{3},
-            new Integer[]{5},
-            new Integer[]{7},
-            new Integer[]{100}
-        );
-    }
-
-    /** Number of partitions. */
-    @Parameterized.Parameter
-    public int parts;
-
+public class LogisticRegressionSGDTrainerTest extends TrainerTest {
     /**
      * Test trainer on classification model y = x.
      */
@@ -65,8 +41,8 @@ public class LogisticRegressionSGDTrainerTest {
     public void trainWithTheLinearlySeparableCase() {
         Map<Integer, double[]> cacheMock = new HashMap<>();
 
-        for (int i = 0; i < data.length; i++)
-            cacheMock.put(i, data[i]);
+        for (int i = 0; i < twoLinearlySeparableClasses.length; i++)
+            cacheMock.put(i, twoLinearlySeparableClasses[i]);
 
         LogisticRegressionSGDTrainer<?> trainer = new LogisticRegressionSGDTrainer<>(new UpdatesStrategy<>(
             new SimpleGDUpdateCalculator().withLearningRate(0.2),
@@ -84,507 +60,4 @@ public class LogisticRegressionSGDTrainerTest {
         TestUtils.assertEquals(0, mdl.apply(VectorUtils.of(100, 10)), PRECISION);
         TestUtils.assertEquals(1, mdl.apply(VectorUtils.of(10, 100)), PRECISION);
     }
-
-    /** Two-easy clustered data by y = x. */
-    private static final double[][] data = {
-        {0.0, -122.69914721554494, -152.90003228835155},
-        {1.0, -988.7803093110984, 39.64498230320555},
-        {1.0, -721.0342526056645, -167.29469954420483},
-        {1.0, 606.0603250738964, 612.4505657575703},
-        {1.0, -435.7428098964267, 749.26660250907},
-        {0.0, 977.0266542119459, 906.2797731011997},
-        {0.0, 442.79191352401017, 99.68443783203702},
-        {1.0, -984.4696576079481, 98.58983213854299},
-        {0.0, 950.3560064579242, -54.087172588871226},
-        {0.0, 989.1247453182418, -942.9228555672748},
-        {1.0, -950.3830359669219, 720.9427578590175},
-        {0.0, -263.7437828854337, -369.67762228969286},
-        {1.0, -837.771820186008, 966.2671117206883},
-        {1.0, -101.63051923258354, 135.30595977925213},
-        {0.0, 927.4068611376827, 552.576689560276},
-        {1.0, 671.674613544031, 867.0342619845135},
-        {0.0, 489.04809639359723, -371.80622025525497},
-        {1.0, -577.8620591314951, -561.9793202960524},
-        {1.0, -628.699903999805, 746.9179933415019},
-        {0.0, 787.7955413710754, 729.8880998762927},
-        {1.0, -160.9905826731191, 597.1342309929371},
-        {1.0, -661.7582546189365, 294.3559610458383},
-        {0.0, 992.067372280372, -586.7840785767917},
-        {0.0, -229.6963941046797, -860.6481903559245},
-        {1.0, -459.91823406828814, 174.31002243199828},
-        {0.0, 132.09417954527203, -203.6015836943012},
-        {0.0, 458.8315635996389, -109.92869423399452},
-        {1.0, 424.63154498678796, 581.7436424491116},
-        {0.0, 606.7777384705123, 382.51034075942744},
-        {1.0, 133.97732363544492, 810.4293150045719},
-        {1.0, -752.3792672455503, 902.3533215842801},
-        {0.0, 124.02578589031486, -242.0045741962906},
-        {0.0, 65.95100120357665, -362.9563512717997},
-        {1.0, -975.7825688109236, -724.6782664271469},
-        {1.0, -885.3333915784285, -166.8285153252507},
-        {1.0, -242.89869955409756, 878.9999767933075},
-        {0.0, 271.2149993049329, -490.0480096390996},
-        {0.0, -74.16302081043352, -824.0859586265949},
-        {1.0, -520.4108075793048, 751.6954919374432},
-        {0.0, 104.03293413801771, -631.0663974778311},
-        {0.0, 179.4274025610996, -610.9764997543232},
-        {1.0, 291.2686412591502, 892.1178988173092},
-        {0.0, 723.1240938478552, -291.3765504086348},
-        {0.0, 12.575218418479949, -307.36975804125973},
-        {1.0, -397.44825972130786, -295.76021536144117},
-        {1.0, -163.90291786947955, 501.6868597449188},
-        {0.0, 513.9232732684154, -287.4072243396091},
-        {1.0, 146.81987289015547, 293.1152654799746},
-        {1.0, -422.734205503476, 154.09536939552663},
-        {0.0, 293.2607563043757, -141.65822134246525},
-        {1.0, -93.46771747630169, 73.91086927080437},
-        {1.0, -972.6525030120272, -867.0819061818511},
-        {1.0, -636.136018043414, 55.4840372628596},
-        {1.0, -821.240801777343, -750.3407912999469},
-        {0.0, 826.9598934792543, -48.17510971836464},
-        {0.0, -737.5399357047692, -834.168742619978},
-        {0.0, 910.2286110591372, -321.2153303241547},
-        {1.0, -539.8385115026349, -204.624635929521},
-        {0.0, 710.9811829617875, 156.53494004963864},
-        {1.0, -576.1327147891295, -255.98030417689222},
-        {0.0, -406.9117225223731, -568.1674835571359},
-        {1.0, 786.4324782672932, 879.9433045727255},
-        {0.0, 655.1507253229393, -931.0320133380443},
-        {1.0, 920.1359556509667, 975.4010808044634},
-        {0.0, 340.9923780361835, -791.6415124130187},
-        {0.0, 789.0326432258107, 101.45600150894029},
-        {0.0, 301.62354598942807, -263.0383267796972},
-        {0.0, -196.75683699829483, -759.6731432356696},
-        {1.0, 104.36756752228234, 362.6645930627608},
-        {0.0, -110.09892045131369, -522.6327938767872},
-        {0.0, 983.058982063912, -853.6685099856713},
-        {0.0, 853.0396544144112, -373.6430440893963},
-        {0.0, 894.5396176478532, -259.3520478430646},
-        {0.0, -59.540445910742505, -405.2785421154832},
-        {1.0, -195.02204474289272, -98.01099074578019},
-        {1.0, -400.33845881394757, 517.4826371806812},
-        {0.0, 998.8721163227847, 658.7589886248159},
-        {1.0, -739.9839264739526, 281.7808456690698},
-        {0.0, 225.2955438875149, -240.13571797647785},
-        {0.0, 415.36363610958847, 119.2467848060553},
-        {1.0, -430.93611072673775, 953.9339020518189},
-        {0.0, 695.641934652828, -613.3163270715312},
-        {1.0, -977.0662561296275, 44.1566618295617},
-        {0.0, 894.0074404584143, 115.97551230630302},
-        {1.0, -256.65810543256225, 121.31432413171797},
-        {1.0, -745.2570475473517, 144.83266177886867},
-        {0.0, 865.266441371979, -329.08860770412593},
-        {1.0, -262.69924145366974, 196.52256942501003},
-        {0.0, 858.8703536921596, -755.3718265129426},
-        {1.0, -620.7574721811682, 744.695289706485},
-        {1.0, 526.9918067706062, 622.6110941283573},
-        {1.0, 30.51838905352247, 451.84360857486945},
-        {1.0, -886.670070825786, 955.5438997547349},
-        {0.0, -419.85446648529296, -904.4363933507589},
-        {1.0, -19.357361515996104, 288.3545217146416},
-        {1.0, 425.807567480902, 617.3859577708511},
-        {1.0, -369.8197242330872, 428.4625522196195},
-        {1.0, -540.2030619980012, 980.1078500916262},
-        {0.0, 963.0216885940265, -999.6718455904652},
-        {0.0, -36.084390168692494, -930.2210871204579},
-        {0.0, 686.7777019875359, 274.083830555807},
-        {1.0, -798.5755214306325, -292.6360310433025},
-        {0.0, -302.49374189510456, -979.2873514693756},
-        {1.0, -473.88156240514184, 290.3700442022921},
-        {1.0, -619.3422333592813, -203.62900604757556},
-        {1.0, -603.8165620304862, 433.7049783716991},
-        {0.0, -394.9003601369652, -423.49571094476414},
-        {0.0, -297.5499912778255, -379.6966117627778},
-        {0.0, 914.6350307682171, 395.0639307730339},
-        {1.0, 302.2432544019764, 420.03068857885364},
-        {1.0, -486.2192439106092, 504.61160963291354},
-        {0.0, -80.9055582464382, -999.3540019713568},
-        {1.0, -808.7735610468485, -600.3003616235419},
-        {1.0, 559.7216432827174, 573.1410775962665},
-        {0.0, 107.25054731907449, 56.68399536280276},
-        {1.0, -986.8173329580039, 955.5975873551458},
-        {0.0, -28.898975148538057, -764.5914349235939},
-        {0.0, 544.5435587517745, 541.7144224905855},
-        {1.0, -733.3388961452514, 995.0625378143936},
-        {0.0, -424.0376248679678, -808.8197992783022},
-        {0.0, 69.10888994619336, -596.3814493832142},
-        {0.0, 668.7563898645246, -309.5338641095864},
-        {1.0, -664.6829023895461, -421.3131122742957},
-        {0.0, 34.30209430645755, -10.50945210920679},
-        {0.0, -370.6335997213754, -510.2102646234516},
-        {1.0, 430.4223842649294, 947.0324231650752},
-        {1.0, -561.4417521638584, 912.0398180862007},
-        {0.0, -529.1099093762112, -787.9426065835444},
-        {0.0, -784.2287272477402, -950.6749150482902},
-        {1.0, -292.2382923363127, 29.73057963193787},
-        {1.0, 543.8216641288004, 574.9668960406921},
-        {0.0, 492.70797586385834, -508.7411915523603},
-        {0.0, 847.4958582226334, 141.27775112134555},
-        {0.0, -294.9950818964355, -539.6512583592041},
-        {1.0, -731.3440778046363, -194.13179207217638},
-        {0.0, -26.21276485761848, -177.1382736912766},
-        {0.0, 169.10051967522577, -877.8835027096119},
-        {0.0, 869.7338657560076, -216.14439990877327},
-        {0.0, 676.9668800100419, 487.3264255975398},
-        {0.0, 340.2086777131092, -483.69798685778176},
-        {0.0, 177.05787101614578, -187.8731928010908},
-        {0.0, 514.0064634256835, -838.309309799528},
-        {1.0, -945.6616134661633, -892.0662652148447},
-        {0.0, 706.7531607568874, 584.875678987067},
-        {0.0, 996.1691889712217, -381.420741757301},
-        {0.0, 846.3827047328193, 138.5937078747695},
-        {1.0, -579.1773394655615, -551.6157981896823},
-        {1.0, -379.8315393213704, 376.240073123181},
-        {0.0, 416.70241675343345, -762.0460887999392},
-        {0.0, 784.4659593773504, -476.3450292459248},
-        {0.0, -328.2495971471759, -797.0282102006712},
-        {1.0, 427.63385513313506, 691.0529822653089},
-        {0.0, 478.22491887051683, 368.08172770775104},
-        {0.0, -194.5486491952804, -635.7562271928532},
-        {1.0, 462.9118544444739, 546.477694721709},
-        {1.0, -364.33646342640543, -16.525517700831642},
-        {1.0, 191.5538518885253, 534.4886561736935},
-        {1.0, 162.29801970257063, 204.07339353277848},
-        {1.0, 359.87375962515307, 510.4390321509045},
-        {0.0, 906.0920707478278, 518.474366833321},
-        {0.0, -23.926514764001354, -545.5535138792807},
-        {1.0, -457.5490330216003, 462.75697632384026},
-        {1.0, 361.19368061986074, 602.0833438729098},
-        {1.0, 240.82404813916537, 903.8580437547587},
-        {0.0, 682.9887385477937, -575.5748494609797},
-        {0.0, -524.9683035626636, -643.4995281011295},
-        {1.0, -868.3907344133812, 687.0334981662659},
-        {0.0, 483.1046447412375, 425.5242965675352},
-        {0.0, 441.7390582141493, -178.6473657093535},
-        {0.0, 857.9901628015248, -725.079106653412},
-        {1.0, 3.9407370946466926, 501.36916187999213},
-        {0.0, 987.6165576421165, -870.7792926909152},
-        {0.0, 38.550394080002434, -316.2460756905849},
-        {1.0, 259.98559430828277, 779.1704474238529},
-        {1.0, -772.0783930084303, 457.81379891960387},
-        {0.0, 965.2460667816263, -900.5906154928432},
-        {0.0, 435.8488975524808, -807.3179393158829},
-        {1.0, -414.9097308847265, 663.2091519493613},
-        {0.0, -692.3369071358595, -853.7674486529854},
-        {1.0, -527.6968945977544, -89.29268231562753},
-        {0.0, 98.58509375449921, -812.2575242800065},
-        {1.0, -246.4858612821199, 690.7736181778389},
-        {0.0, 306.0413673433336, 50.36342267895475},
-        {0.0, -326.3755954952927, -630.9271581822045},
-        {0.0, 435.3759701541835, -478.72141764190417},
-        {0.0, 150.07627192243012, -126.16495181072969},
-        {0.0, 999.2382522208045, 293.8336213483592},
-        {1.0, -970.7818229850416, 559.8116781984274},
-        {0.0, 321.62133209742956, -446.07065722044115},
-        {1.0, 387.61470906465297, 809.9877801153038},
-        {1.0, 375.48380231362376, 548.1340438996276},
-        {0.0, 198.31962497327982, -841.3407638914643},
-        {0.0, -59.75027524961797, -196.91881794207666},
-        {0.0, 539.4390329297466, 265.73233936446013},
-        {1.0, 161.7769611006779, 420.4911194344545},
-        {1.0, -422.73262266569805, 305.27632230640575},
-        {0.0, 419.7041783295376, 384.4277361814418},
-        {1.0, -384.80122335064925, 128.84723939702212},
-        {0.0, 345.8732451410485, -634.6766931661393},
-        {1.0, -753.0957875425104, 162.043321600848},
-        {1.0, -721.0825943433963, -647.1437151757809},
-        {0.0, 737.8179495142201, -612.9000146979762},
-        {0.0, 165.62609685662937, -209.04556534374638},
-        {1.0, 211.75025757991534, 762.4363190775396},
-        {0.0, -282.0707259050812, -631.5669067165459},
-        {0.0, -10.649387489441551, -11.742073063187377},
-        {0.0, 532.2273317939553, -714.4637938741703},
-        {0.0, 851.6255007653094, -428.168617931829},
-        {0.0, -650.2303513768155, -701.0819971407498},
-        {0.0, 486.19072881419584, 17.642342348021202},
-        {0.0, 937.5878660613639, 253.91073899684488},
-        {1.0, -481.7837261941776, 386.0515070365086},
-        {1.0, 898.8591491398315, 960.3282479515362},
-        {1.0, -795.2119099095994, -52.442255260638944},
-        {1.0, -832.14760576095, 406.48368080778823},
-        {1.0, 317.3610961002403, 475.88090137988934},
-        {1.0, -543.9941239514503, 937.9571974443777},
-        {1.0, -737.7149868841586, 412.02870959820666},
-        {1.0, -86.04799530647608, 764.2717139104996},
-        {1.0, -908.3441434769735, -52.62148904481751},
-        {1.0, -558.4878652128368, 975.5017115797407},
-        {1.0, -120.28961819893993, 58.60059810912276},
-        {0.0, 797.7665926374921, -530.0884822652556},
-        {0.0, -248.62486746176887, -983.5555931167586},
-        {0.0, 910.1931415438364, 35.953135142478914},
-        {1.0, -304.741023136228, 253.0138864886694},
-        {0.0, -510.13133519018925, -642.3600729680307},
-        {0.0, 988.5683650098642, -751.2030447890847},
-        {1.0, -118.0142080751416, 352.20209758019996},
-        {0.0, -638.757222741898, -685.6631975937353},
-        {0.0, 759.5622347453971, -722.2769348273996},
-        {0.0, -740.3498419247273, -974.2677201928796},
-        {0.0, -776.6102763008262, -993.7697826767383},
-        {1.0, -895.9448277148887, -462.29125820523006},
-        {0.0, -311.8810163384071, -318.9742942085709},
-        {0.0, 368.78035230644787, -273.65009131252566},
-        {0.0, 731.1488644867686, -184.2725009666142},
-        {1.0, 240.0262332913362, 544.8792933528591},
-        {1.0, -129.8786600652611, 122.64122390591797},
-        {1.0, -998.8693504661202, -989.3959455521401},
-        {0.0, 358.9021702584721, -372.46195332982563},
-        {0.0, 423.66170839399, -3.6733713507491075},
-        {0.0, 320.08527272511014, -267.49487239617406},
-        {1.0, 628.8557340365153, 716.1736088420723},
-        {1.0, 87.0852622927755, 191.08205494997515},
-        {0.0, -163.5535634273158, -401.43333064263857},
-        {1.0, 241.57291015127043, 354.07473809573935},
-        {0.0, 425.42982178930424, -659.6389818980119},
-        {1.0, -513.057622632338, -150.48805414197307},
-        {0.0, 435.2888705572377, -500.4699931158425},
-        {1.0, -761.2341202466506, 919.1637075257438},
-        {1.0, -254.8539665845866, 711.5522826694619},
-        {0.0, -350.2587997576785, -911.7842377194485},
-        {0.0, 588.5547568621123, -16.003674634160916},
-        {0.0, -557.7880688291352, -939.7740734026603},
-        {0.0, 683.6988697659988, -285.8831419034458},
-        {0.0, 782.8461154585116, 426.91516285206694},
-        {1.0, -792.3388875152918, 361.1342300030676},
-        {1.0, -673.792921360787, 820.8934158286147},
-        {1.0, -15.357504282120772, 15.275909249335541},
-        {0.0, -99.22050275699814, -249.077767711845},
-        {1.0, -820.111231678807, -320.1107983145504},
-        {0.0, 911.7878651586336, 825.2998851049153},
-        {1.0, -750.2941326911656, -629.1546336560141},
-        {1.0, -890.6374102685097, -804.5407239545832},
-        {1.0, -204.75148861468108, 722.1116624961337},
-        {0.0, 519.1714356909579, 154.07772492651725},
-        {0.0, 982.2450336212896, 897.8824490832485},
-        {0.0, 554.4793545664838, 335.7541373769475},
-        {1.0, -339.90247025178235, 47.02715071976445},
-        {0.0, 901.2543768759774, -662.3275399668249},
-        {1.0, -942.3762411246095, -875.0025895092708},
-        {0.0, 418.20256050104604, -414.3102074305251},
-        {0.0, 625.0294460702908, -625.6315794655841},
-        {1.0, -449.74570685873516, 937.185777575773},
-        {0.0, 508.2386960118979, 454.0962431757914},
-        {1.0, 331.4089009636193, 589.2741722009719},
-        {1.0, 99.06469391982864, 187.0394494146019},
-        {1.0, -982.3370248476699, 322.0973186273661},
-        {1.0, 548.6443983489316, 708.7265431968447},
-        {0.0, 918.9454013804204, -383.8602043941679},
-        {1.0, 47.025960736300476, 171.219298464468},
-        {0.0, 378.2597384891858, 163.1492885941102},
-        {0.0, 438.65288112462554, -139.6734662005057},
-        {1.0, -831.8875659762939, 892.6667556591465},
-        {0.0, 883.0433572247841, -405.08376291753257},
-        {0.0, 885.9349479866808, -577.4873262774219},
-        {1.0, -614.7099535083557, -133.06983968843338},
-        {0.0, 111.7257364798395, -585.9016094589116},
-        {1.0, 453.9214560104581, 999.4093349063546},
-        {1.0, -660.6080448479984, -558.4295455433598},
-        {0.0, -466.8209751830958, -591.196870091049},
-        {0.0, -964.7665601618734, -997.9800903796079},
-        {0.0, -236.07763234295055, -450.41129146522917},
-        {0.0, -621.6876241277605, -797.4500041783042},
-        {0.0, -773.3591978507126, -890.0043590247606},
-        {1.0, -41.04699663875965, 822.3779367276668},
-        {0.0, -88.10853803965915, -192.37350885363378},
-        {0.0, 663.981740050287, -508.81572667480236},
-        {0.0, 15.59472374839936, -806.7541810675616},
-        {1.0, -892.7104844234832, -708.5235867565298},
-        {1.0, -484.65491520217245, 386.6430150137869},
-        {0.0, 865.0610549279427, 615.8811284084713},
-        {1.0, -824.4355093837889, 655.3234320109748},
-        {1.0, -274.68139814419976, -239.53727115479273},
-        {0.0, -86.4277464637313, -881.0777192437689},
-        {1.0, -581.4932661460668, 769.3538369247574},
-        {0.0, -432.5850223289913, -577.4260081674186},
-        {1.0, 166.76522990130684, 582.4331818363789},
-        {0.0, 396.8182460459341, 248.34183939490367},
-        {1.0, -509.8701926143476, 368.8796357552451},
-        {1.0, -482.54152901054886, -248.83959837521047},
-        {1.0, -300.50297994358345, 742.4139758199028},
-        {0.0, 163.28493788474384, -61.41706872692157},
-        {0.0, -399.2277405988791, -930.6519043114885},
-        {0.0, 44.13900477801826, -571.5314250642764},
-        {0.0, 457.8794897532496, -505.99693186447195},
-        {0.0, 16.85880382123935, -451.1811783607169},
-        {1.0, -743.4540696447744, 325.39937301862096},
-        {1.0, 57.40459247973081, 106.58399169789641},
-        {1.0, 183.98880310846016, 499.74779967287395},
-        {1.0, 567.0903172389608, 820.4986606446041},
-        {0.0, 672.4806526088855, 300.601012280614},
-        {0.0, -343.8894522407976, -761.4942297431235},
-        {0.0, 870.247864223385, -168.14608036197296},
-        {1.0, 593.005455426467, 673.1630290763387},
-        {0.0, -625.9494316959813, -983.6968015830237},
-        {1.0, 494.1754094118269, 992.2691899024903},
-        {0.0, 61.401789304312615, -773.2837841463802},
-        {1.0, -194.76742246565573, 69.77988116139159},
-        {0.0, 206.82364861578685, 121.15474801344544},
-        {1.0, -265.964495521001, 50.790074285276205},
-        {0.0, 818.3873132702915, 36.49793444927877},
-        {0.0, 99.81409878465752, -628.0274914181116},
-        {0.0, 464.149315901346, -321.29715928735277},
-        {1.0, -164.52462729937565, 952.4896905712137},
-        {0.0, -63.17364851415209, -149.49056773721736},
-        {0.0, 882.9288293898815, 171.00117804059573},
-        {0.0, 473.3733180102365, -689.3426862684687},
-        {0.0, 165.7220875180078, -354.71003889056044},
-        {0.0, 525.5517697849327, 415.84107073078167},
-        {0.0, -38.184721358457864, -99.36030799911896},
-        {0.0, 242.96729902384163, -156.16029387422054},
-        {0.0, 448.4711090805122, -495.01683482080705},
-        {1.0, -80.15226220702493, 970.6850105496733},
-        {0.0, 870.3328249998483, 583.0363909361256},
-        {0.0, -238.61798549246464, -430.95739845768026},
-        {0.0, -153.01230031899092, -482.12077718764306},
-        {1.0, -118.06183953458049, 40.44154430898425},
-        {1.0, -876.8968143885145, -370.6419068924105},
-        {0.0, 989.8165746071368, -943.0636134966381},
-        {0.0, 448.68476431428917, 44.44832374987436},
-        {0.0, -5.562631397638029, -594.7883897866259},
-        {0.0, 880.7175397337289, 786.6444839355895},
-        {0.0, 476.3278235630439, -756.8025350513306},
-        {0.0, -209.1261948306602, -366.9709734757247},
-        {1.0, -1.5342655753494228, 295.69953419777266},
-        {1.0, 98.88194946977887, 709.984198980128},
-        {1.0, -102.4522435336255, 348.55854643990347},
-        {1.0, 431.6422144084929, 488.26608578711966},
-        {1.0, -629.5648689407153, -389.98821373225144},
-        {1.0, -655.6263155228037, 89.12505314113082},
-        {0.0, -201.6475575882739, -902.9470477574147},
-        {1.0, -342.30143560116915, 157.21169053018912},
-        {1.0, -671.4797028289656, -49.48397951858112},
-        {1.0, -993.3541982679827, 428.50119148048657},
-        {0.0, 158.95824836793054, 115.93705315336206},
-        {1.0, -858.292999815246, 946.8912002937116},
-        {1.0, -223.10861890967476, 190.7507270694814},
-        {0.0, -147.9091707330915, -899.2785339400244},
-        {0.0, 254.55648822491457, -260.9331332388332},
-        {0.0, 560.3172529225217, 388.76836664538814},
-        {0.0, 924.1007767093995, 56.69156104001263},
-        {0.0, 62.42705110549082, -888.0360838024912},
-        {0.0, 222.43761905783595, 88.18795871018938},
-        {0.0, 489.8756173625022, 421.3474970424486},
-        {0.0, 246.6646015601891, -506.3175818566548},
-        {0.0, -620.5001534479718, -774.7836865370457},
-        {1.0, -654.0153133260937, -369.1547696738236},
-        {0.0, 853.1429595371762, -87.56985188355861},
-        {1.0, -226.84561483455388, 122.80144293902458},
-        {1.0, 335.09779003775316, 731.0032200516428},
-        {1.0, 87.90214612318391, 724.8989520503376},
-        {0.0, -51.792728592205776, -298.0103777307395},
-        {1.0, -421.181682827191, 41.01565470282776},
-        {1.0, -626.6392286104665, 227.98017875883284},
-        {1.0, -839.0341042344045, 990.7893877153003},
-        {1.0, -9.321936022159207, 125.24249479969853},
-        {0.0, 665.2916192497585, 314.9312297793483},
-        {1.0, -236.71130814979108, 41.56269468081973},
-        {1.0, -695.4935496704909, -364.376100277162},
-        {0.0, 60.90303121087936, -525.9732822401365},
-        {1.0, -740.9211189318623, 328.1577766746841},
-        {0.0, 636.7728693761635, 231.63887313030887},
-        {0.0, 783.8640093145868, -86.94016828207737},
-        {1.0, -122.79445443476675, 446.8427679254348},
-        {0.0, -599.127065456006, -641.9946421169902},
-        {0.0, -133.3932116798295, -715.8087793479069},
-        {0.0, 868.1768857382554, -356.8832640029416},
-        {1.0, -729.5079555062296, 48.18869346933934},
-        {1.0, -323.311327276945, 51.37289795053448},
-        {1.0, -863.9094602749768, -526.3307161874084},
-        {0.0, -172.237643059304, -545.395840196842},
-        {1.0, 379.0803154405653, 860.9286051762328},
-        {0.0, 646.3490077056538, 221.13771257535495},
-        {1.0, -493.2329575593668, 938.8602740452263},
-        {0.0, 852.1508064390962, 186.42129731281898},
-        {0.0, -105.17633183875978, -819.8477185986328},
-        {0.0, 794.7790444633961, 225.19911969860573},
-        {0.0, 306.4485552684148, 290.3991023596727},
-        {1.0, -348.52545404552563, -302.8538669615166},
-        {1.0, -621.5896829696857, -586.764214213187},
-        {0.0, -360.9052184666539, -501.2314262330038},
-        {0.0, 512.0475423578778, -968.4211685736286},
-        {0.0, -1.0553261239787162, -649.1131987920394},
-        {1.0, -353.0059560079317, -343.82940709059096},
-        {0.0, 281.71038662642286, -536.6960537047482},
-        {1.0, -919.2355704939898, 782.9875939766282},
-        {1.0, -554.7648476025646, 670.76664941987},
-        {0.0, 287.54041983444336, 106.2628262971964},
-        {1.0, -71.36414070058743, 481.00905876949264},
-        {1.0, -525.4581932812421, 507.16990298296923},
-        {0.0, 510.1084615227803, -813.3443471544821},
-        {0.0, -515.8000398448883, -551.1523846072581},
-        {1.0, -941.5905835281701, 178.53493537516124},
-        {1.0, -826.4320007540575, -391.32308974320074},
-        {1.0, -362.25207668798646, 711.1776477575349},
-        {1.0, -363.13146140965796, 58.76850122459791},
-        {1.0, -637.0939514034111, -57.18171960880602},
-        {1.0, 811.8537434287423, 893.8406118576338},
-        {1.0, -351.36128471993413, -164.8367432830371},
-        {0.0, -625.8073644486308, -938.5091097468568},
-        {0.0, 131.36904305993585, 59.945922200265386},
-        {1.0, 300.49666138667953, 544.089396622054},
-        {1.0, 150.9533638033147, 943.667562848439},
-        {1.0, -232.3556550990304, 976.0470122102599},
-        {1.0, 135.8097187722467, 262.21166985817695},
-        {0.0, -97.51353115825805, -890.6273287611524},
-        {1.0, -711.4020131465077, -20.13806627790268},
-        {0.0, 917.1543030685937, -872.6562190191934},
-        {1.0, -657.7632592299774, -596.4956657628013},
-        {0.0, 806.7273372492091, 154.3973882475018},
-        {0.0, 371.7932221354017, -847.5721372522485},
-        {0.0, 887.0251089691258, -306.6059397900773},
-        {1.0, -171.52557116367404, 819.6507572581761},
-        {0.0, 632.2374116222845, -635.8014704304069},
-        {0.0, -213.33363068356653, -639.038384943213},
-        {0.0, 737.7847710201636, -843.291366957395},
-        {0.0, -430.7114667797465, -665.7014140302028},
-        {0.0, 18.317432837854085, -309.1307864153605},
-        {0.0, 689.3196508440624, 398.22692583132357},
-        {0.0, 908.6965655126414, -321.7431267700932},
-        {0.0, 604.2361606207025, -174.1208906780612},
-        {1.0, -816.014328616853, -468.5728222442267},
-        {1.0, -124.50677921712554, 439.4225345583168},
-        {0.0, -736.4729915358428, -745.435394454091},
-        {1.0, -201.1314081356761, 132.070557003796},
-        {1.0, -538.2469045343253, 719.2630473774586},
-        {1.0, -579.3039091203984, 961.7644587928542},
-        {1.0, -131.07569768983058, -14.067659190625022},
-        {1.0, -961.9324831150435, 815.7775199747161},
-        {0.0, 959.0805916122792, 210.22031178108682},
-        {0.0, 537.3004634155134, -821.1232504829824},
-        {1.0, -525.577776451393, 523.8546325250404},
-        {1.0, -490.37425007561785, 613.9247103792861},
-        {1.0, 725.2941641152454, 924.7691776631311},
-        {0.0, 850.5191959199387, -911.7156754307339},
-        {1.0, -535.3827552133765, -256.1333041657481},
-        {1.0, 93.24441210512305, 980.899958839474},
-        {1.0, 125.58210878499744, 489.9200659506546},
-        {1.0, -265.0907509361897, -181.36232727265053},
-        {1.0, -805.0528978104943, -774.3428711441273},
-        {0.0, 299.481029365769, 274.2467784888322},
-        {1.0, -872.6432839751412, -724.9692038478101},
-        {0.0, -327.77109720806027, -346.06090524099113},
-        {0.0, -769.9407295518204, -947.4499512111647},
-        {0.0, 708.176001237056, -701.9900242821255},
-        {0.0, 429.7900423607498, -767.8607100772805},
-        {0.0, 514.9666605063433, -252.09527799878242},
-        {1.0, -392.6943024744394, 943.3642876383242},
-        {0.0, -171.97676164837765, -964.9749845719992},
-        {0.0, 25.3949751703301, -761.3459408840288},
-        {0.0, 327.0516125752938, -81.26274312696592},
-        {0.0, -926.4851014957853, -970.9563176084357},
-        {1.0, -985.2416286372801, -758.6127879964147},
-        {0.0, 338.7854869375187, -231.37122411100802},
-        {1.0, -995.9157184785086, -310.8674450540059},
-        {0.0, 485.52790893379097, 7.909018196822899},
-        {1.0, -289.76601009744377, -93.43411467378803},
-        {1.0, -352.91681813664957, 970.6609344632727},
-        {1.0, -634.2596635738871, 478.54324561131875},
-        {1.0, -496.623286353002, 526.7778661797483},
-        {0.0, 837.0404771301767, 671.1823960639354},
-        {0.0, -284.5931069950618, -893.2503900000672},
-        {0.0, 739.6925158457948, -572.886151546864},
-        {1.0, 505.37418939555437, 914.4939776238757},
-        {0.0, 65.79978723030536, -59.26282586191303},
-        {0.0, 775.1318885055389, -698.3367782064498},
-        {1.0, -871.3166585822554, -351.74555670546727}
-    };
 }
